fix for int-overflows with huge queues
fix for when nzbget couldn't handle huge queues due to the limitation of int32-values and ended up crashing

+++ daemon/extension/CommandScript.h
@@ -54,7 +54,7 @@ public:
 private:
 	MessageList m_messages;
 	Mutex m_logMutex;
-	int m_idMessageGen;
+	int64 m_idMessageGen;
 	int m_idScriptGen;
 };
